moving to jooq and compose.wasm 1.5.10-beta02

+++ backend/src/main/kotlin/dev/silas/domain/Link.kt
@@ -1,6 +1,5 @@
 package dev.silas.domain
 
-import java.math.BigInteger
 import java.time.OffsetDateTime
 import java.util.UUID
 
@@ -8,6 +7,6 @@ data class Link(
     val id: UUID,
     val shortUrl: String,
     val fullUrl: String,
-    val hits: BigInteger,
+    val hits: Int,
     val createdAt: OffsetDateTime
 )
